Now mapping cfitsio TLONG to C long to fix FITS binary imports on 64 bit archs.
In retrospect, this may look trivial :-)

+++ gavo/grammars/directgrammar.py
@@ -534,7 +534,7 @@ class FITSCodeGenerator(_CodeGenerator):
 		"B": ("TBYTE", "char"),
 		"A": ("TSTRING", "char *"),
 		"I": ("TSHORT", "short"),
-		"J": ("TLONG", "int"),
+		"J": ("TLONG", "long"),
 		"K": ("TLONGLONG", "long long"),
 		"E": ("TFLOAT", "float"),
 		"D": ("TDOUBLE", "double")}
